requirements.txt
Flask
pycryptodome
requests
protobuf
PyJWT